requirements.txt
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
pdfminer.six==20211012
pycparser==2.21
